Fixed a few warnings.

+++ src/preprocessor.h
@@ -736,7 +736,7 @@ static ptrdiff_t calc1(struct cpp *this, WCHAR *data, ptrdiff_t len,
   return pos;
 }
 
-static int calc(struct cpp *this, WCHAR *data, ptrdiff_t len, ptrdiff_t tmp)
+static ptrdiff_t calc(struct cpp *this, WCHAR *data, ptrdiff_t len, ptrdiff_t tmp)
 {
   JMP_BUF recovery;
   INT32 pos;
@@ -772,12 +772,12 @@ static int calc(struct cpp *this, WCHAR *data, ptrdiff_t len, ptrdiff_t tmp)
   return pos;
 }
 
-static INT32 lower_cpp(struct cpp *this,
-		       WCHAR *data,
-		       ptrdiff_t len,
-		       int flags,
-		       int auto_convert,
-		       struct pike_string *charset)
+static ptrdiff_t lower_cpp(struct cpp *this,
+			   WCHAR *data,
+			   ptrdiff_t len,
+			   int flags,
+			   int auto_convert,
+			   struct pike_string *charset)
 {
   ptrdiff_t pos, tmp, e;
   int tmp2;
